refactor(backend): merge user to misc resolver
