+++ apps/backend/src/main.rs
@@ -20,6 +20,7 @@ use axum::{
 use config::AppConfig;
 use dotenvy::dotenv;
 use http::header::AUTHORIZATION;
+use misc::resolver::COOKIE_NAME;
 use rust_embed::RustEmbed;
 use sea_orm::{Database, DatabaseConnection};
 use sea_orm_migration::MigratorTrait;
@@ -46,7 +47,6 @@ use crate::{
     config::get_app_config,
     graphql::{get_schema, GraphqlSchema},
     migrator::Migrator,
-    users::resolver::COOKIE_NAME,
     utils::create_app_services,
 };
 
@@ -64,7 +64,6 @@ mod movies;
 mod podcasts;
 mod shows;
 mod traits;
-mod users;
 mod utils;
 mod video_games;
 
@@ -172,10 +171,10 @@ async fn main() -> Result<()> {
     let importer_service_1 = app_services.importer_service.clone();
     let importer_service_2 = app_services.importer_service.clone();
     let media_service_1 = app_services.media_service.clone();
-    let users_service_1 = app_services.users_service.clone();
-    let users_service_2 = app_services.users_service.clone();
-    let users_service_3 = app_services.users_service.clone();
     let misc_service_1 = app_services.misc_service.clone();
+    let misc_service_2 = app_services.misc_service.clone();
+    let misc_service_3 = app_services.misc_service.clone();
+    let misc_service_4 = app_services.misc_service.clone();
     let monitor = async {
         let mn = Monitor::new()
             // cron jobs
@@ -195,7 +194,7 @@ async fn main() -> Result<()> {
                     )
                     .layer(ApalisTraceLayer::new())
                     .layer(ApalisExtension(app_services.media_service.clone()))
-                    .layer(ApalisExtension(users_service_1.clone()))
+                    .layer(ApalisExtension(misc_service_1.clone()))
                     .build_fn(general_user_cleanup)
             })
             .register_with_count(1, move |c| {
@@ -223,7 +222,7 @@ async fn main() -> Result<()> {
             .register_with_count(1, move |c| {
                 WorkerBuilder::new(format!("user_created_job-{c}"))
                     .layer(ApalisTraceLayer::new())
-                    .layer(ApalisExtension(users_service_2.clone()))
+                    .layer(ApalisExtension(misc_service_2.clone()))
                     .with_storage(user_created_job_storage.clone())
                     .build_fn(user_created_job)
             })
@@ -237,7 +236,7 @@ async fn main() -> Result<()> {
             .register_with_count(1, move |c| {
                 WorkerBuilder::new(format!("recalculate_user_summary_job-{c}"))
                     .layer(ApalisTraceLayer::new())
-                    .layer(ApalisExtension(users_service_3.clone()))
+                    .layer(ApalisExtension(misc_service_3.clone()))
                     .with_storage(recalculate_user_summary_job_storage.clone())
                     .build_fn(recalculate_user_summary_job)
             })
@@ -245,7 +244,7 @@ async fn main() -> Result<()> {
                 WorkerBuilder::new(format!("update_metadata_job-{c}"))
                     .layer(ApalisTraceLayer::new())
                     .layer(ApalisRateLimitLayer::new(3, Duration::new(5, 0)))
-                    .layer(ApalisExtension(misc_service_1.clone()))
+                    .layer(ApalisExtension(misc_service_4.clone()))
                     .with_storage(update_metadata_job_storage.clone())
                     .build_fn(update_metadata_job)
             })
