--- apps/backend/src/users/mod.rs
@@ -1,1 +1,0 @@
-pub mod resolver;
